medicine_model_curasense/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-generativeai
groq
tavily-python
chromadb
sentence-transformers
beautifulsoup4
requests
python-multipart
